avoid recursive instantiation loops
Filling the cache with None before instantiating/interpreting the new
expression avoids it being recursively instantiated. The caller should
be able to handle a None value instead of a proper expression, which
seems to be the case (the resulting co_constraint will just be None, a
case which is handled).